Add --force-optimization-level option to Tools/Scripts/set-webkit-configuration. https://bugs.webkit.org/show_bug.cgi?id=205787
Reviewed by Saam Barati.

Usage:
    --force-optimization-level=<opt> Force optimization: O3, O2, O1, O0, Os, Ofast, Og, or none

This can be used to force debug builds to be built with a higher level optimization
so that tests can run to completion faster.

It can also be useful as a simple way to force release builds to be built with
different optimization levels for performance comparison.

Setting --force-optimization-level=none restores the default optimization levels.
Of course, the build targets need to be rebuilt for this to take effect.

* Scripts/set-webkit-configuration:
* Scripts/webkitdirs.pm:
(determineForceOptimizationLevel):
(forceOptimizationLevel):
(XcodeOptions):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@254080 268f45cc-cd09-0410-ab3c-d52691b4dbfc