[LFC][IFC] Visually collapse hanging pre-wrap content. https://bugs.webkit.org/show_bug.cgi?id=206133 <rdar://problem/58505750>
Reviewed by Antti Koivisto.

This change is to comply with other rendering engines when it comes to visually collapsing hanging pre-wrap content.

https://www.w3.org/TR/css-text-3/#white-space-phase-2
"If white-space is set to pre-wrap, the UA must (unconditionally) hang this sequence, unless the sequence
is followed by a forced line break, in which case it must conditionally hang the sequence is instead.
It ___may___ also visually collapse the character advance widths of any that would otherwise overflow."

* layout/inlineformatting/InlineLineBuilder.cpp:
(WebCore::Layout::LineBuilder::close):
(WebCore::Layout::LineBuilder::visuallyCollapsePreWrapOverflowContent):
* layout/inlineformatting/InlineLineBuilder.h:
(WebCore::Layout::LineBuilder::InlineItemRun::adjustLogicalWidth):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@254402 268f45cc-cd09-0410-ab3c-d52691b4dbfc