CacheStorage::Engine::clearCachesForOriginFromDisk ASSERT is buggy https://bugs.webkit.org/show_bug.cgi?id=206057 <rdar://problem/57762994>
Reviewed by John Wilander.

The ASSERT is only valid if the engine is initialized.
It is not needed to initialize it if the plan is to remove all disk files.
Instead, update ASSERT to check that either m_salt is not there or the salt is there and the path is as expected.

* NetworkProcess/cache/CacheStorageEngine.cpp:
(WebKit::CacheStorage::Engine::clearCachesForOriginFromDirectories):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@254501 268f45cc-cd09-0410-ab3c-d52691b4dbfc