Web Inspector: show the device settings menu when Web Inspector's debug mode is enabled https://bugs.webkit.org/show_bug.cgi?id=206311
Reviewed by Brian Burg.

This only has an effect when Web Inspector is opened and the debug mode is enabled. If so,
the device settings toolbar icon will appear regardless of whether the inspected target is
a remote device or not. Disabling debug mode will not immediately remove the device settings
toolbar icon, but it will not reappear when Web Inspector is next opened (to allow for taking
screenshots and other such things).

* UserInterface/Base/Main.js:
(WI.contentLoaded):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@254727 268f45cc-cd09-0410-ab3c-d52691b4dbfc