Fix SDK availability macros https://bugs.webkit.org/show_bug.cgi?id=206463 <rdar://problem/58581906>
Reviewed by Timothy Hatcher.

Bug 197841 updated WebKit.xcconfig with regards to when it runs a
header post-processing script. This post-processing script grovels
over our exported headers, changing macros like WK_API_AVAILABLE to
the standard API_AVAILABLE. Because of the change in Bug 197841, the
script was no longer being run when preparing macOS 10.15 SDKs, which
was breaking those SDKs. Fix this by reverting the change, allowing
the macros to be replaced for macOS 10.15.

No new tests - no added or changed functionality.

* Configurations/WebKit.xcconfig:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@254876 268f45cc-cd09-0410-ab3c-d52691b4dbfc