Remove some unneeded definitions from Platform.h family headers https://bugs.webkit.org/show_bug.cgi?id=206642
Reviewed by Eric Carlson.

Source/WebCore:

* platform/graphics/avfoundation/objc/InbandTextTrackPrivateAVFObjC.h: Removed
checks of HAVE(AVFOUNDATION_MEDIA_SELECTION_GROUP) and
HAVE(AVFOUNDATION_VIDEO_OUTPUT) because both are always 1 on Cocoa platforms,
and this is Cocoa-only source code.
* platform/graphics/avfoundation/objc/InbandTextTrackPrivateAVFObjC.mm: Ditto.
* platform/graphics/avfoundation/objc/MediaPlayerPrivateAVFoundationObjC.h: Ditto.
* platform/graphics/avfoundation/objc/MediaPlayerPrivateAVFoundationObjC.mm: Ditto.

Source/WebCore/PAL:

* pal/spi/cf/CFNetworkSPI.h: Remove
HAVE(CFNETWORK_WITH_CONTENT_ENCODING_SNIFFING_OVERRIDE) because it was
always 1 on Cocoa platforms, and was guarding a definition that we
use unconditionally in Objective-C Cocoa-only code, and there is no
harm of defining it on Windows as long as we don't use it.

* pal/spi/cocoa/PassKitSPI.h: Remove HAVE(PASSKIT_API_TYPE) because
it was always 1 on Cocoa platforms and this is a Cocoa-only header.

Source/WebKit:

* Platform/cocoa/PaymentAuthorizationViewController.mm:
(-[WKPaymentAuthorizationViewControllerDelegate _getPaymentServicesMerchantURL:]):
Removed HAVE(PASSKIT_API_TYPE) because it's always 1 on Cocoa platforms and this
is Cocoa-only code.
* Shared/ApplePay/cocoa/WebPaymentCoordinatorProxyCocoa.mm:
(WebKit::toAPIType): Ditto.
(WebKit::WebPaymentCoordinatorProxy::platformPaymentRequest): Ditto.

Source/WTF:

* wtf/PlatformEnable.h:
Removed ENABLE_RESPECT_EXIF_ORIENTATION because it was unused.

* wtf/PlatformHave.h:
Removed HAVE_PASSKIT_API_TYPE because it is now unused.
Removed HAVE_MERGESORT because it was unused.
Removed HAVE_AVFOUNDATION_VIDEO_OUTPUT because it is now unused.
Removed HAVE_AVFOUNDATION_MEDIA_SELECTION_GROUP because it is now unused.
Removed HAVE_CFNETWORK_WITH_CONTENT_ENCODING_SNIFFING_OVERRIDE because it was unused.

* wtf/PlatformUse.h:
Removed USE_ARENA_ALLOC_ALIGNMENT_INTEGER because it was unused.
Removed USE_NETWORK_CFDATA_ARRAY_CALLBACK because it is now unused.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@255049 268f45cc-cd09-0410-ab3c-d52691b4dbfc