Need a way to mark a DOM attribute as runtime conditionally read-write https://bugs.webkit.org/show_bug.cgi?id=206836
Patch by Sam Weinig <weinig@apple.com> on 2020-01-27
Reviewed by Chris Dumez.

Adds a new IDL extended attribute, RuntimeConditionallyReadWrite, that works
like EnableAtRuntime, but instead allows the associated attribute to be
converted from read-only to read-write based on the value of a RuntimeEnabledFeature.

* bindings/scripts/CodeGeneratorJS.pm:
(GeneratePropertiesHashTable):
Add check that both runtime enabling and runtime conditional read-write aren't
both specified. For now, that is not supported.

(GenerateRuntimeEnableConditionalString):
Add support for generating the conditional string for RuntimeConditionallyReadWrite.

(GenerateImplementation):
Implement the feature by piggy-backing off the existing runtime enabling, which
deletes the property if the feature is not enabled. For this, if the feature isn't
enabled, we delete the property, and then add it back readonly. In the future, we
should move off of this add-then-delete model, but that is tracked via http://webkit.org/b/206837.

* bindings/scripts/IDLAttributes.json:
Add new extended attribute.

* bindings/scripts/test/JS/JSTestObj.cpp:
(WebCore::JSTestObjPrototype::finishCreation):
(WebCore::jsTestObjRuntimeConditionallyReadWriteAttributeGetter):
(WebCore::jsTestObjRuntimeConditionallyReadWriteAttribute):
(WebCore::setJSTestObjRuntimeConditionallyReadWriteAttributeSetter):
(WebCore::setJSTestObjRuntimeConditionallyReadWriteAttribute):
* bindings/scripts/test/TestObj.idl:
Add new test with result.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@255224 268f45cc-cd09-0410-ab3c-d52691b4dbfc