[ macOS wk2 ] http/tests/media/media-stream/get-display-media-prompt.html is flaky failure https://bugs.webkit.org/show_bug.cgi?id=206958 <rdar://problem/59003765>
Reviewed by Eric Carlson.

We added a rule to only allow one getDisplayMedia call per gesture.
For that reason, we were storing a pointer to the gesture event and
resetting in case the current gesture event was equal to the stored pointer.
Instead, store a WeakPtr which ensures that if the previous event is destroyed,
the weak pointer will be null and so will not have the same value as the new gesture event.
Covered by existing tests no longer being flaky.

* Modules/mediastream/MediaDevices.cpp:
(WebCore::MediaDevices::computeUserGesturePriviledge):
* Modules/mediastream/MediaDevices.h:
* dom/UserGestureIndicator.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@255582 268f45cc-cd09-0410-ab3c-d52691b4dbfc