Source/WebKit: Pages that trigger a redirect will sometimes be left blank https://bugs.webkit.org/show_bug.cgi?id=207614 rdar://problem/59077740
Patch by Simon Fraser <simon.fraser@apple.com> on 2020-02-12
Reviewed by Tim Horton.

TiledCoreAnimationDrawingArea::setRootCompositingGraphicsLayer() can be called when the layer tree
is frozen, in which case we stash away the layer in m_pendingRootLayer to be parented later at flush
time. However, this sequence of calls had a bug:

setRootCompositingGraphicsLayer() when frozen
    -> stash in m_pendingRootLayer
setRootCompositingGraphicsLayer() when not frozen
    -> set the root layer
flushLayers()
    -> set the root layer to the (old) m_pendingRootLayer

So we need to clear m_pendingRootLayer at step 2.

Very timing dependent, hard to test.

* WebProcess/WebPage/mac/TiledCoreAnimationDrawingArea.mm:
(WebKit::TiledCoreAnimationDrawingArea::setRootCompositingLayer):

LayoutTests:
[iOS] Deny mach lookup access to analytics service in the WebContent process
https://bugs.webkit.org/show_bug.cgi?id=207482

Reviewed by Darin Adler.

* fast/sandbox/ios/sandbox-mach-lookup-expected.txt:
* fast/sandbox/ios/sandbox-mach-lookup.html:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@256455 268f45cc-cd09-0410-ab3c-d52691b4dbfc