[iOS] Check if PIP is supported in the UI process https://bugs.webkit.org/show_bug.cgi?id=207406
Reviewed by Brent Fulgham.

Source/WebCore:

This is currently being checked in the WebProcess, but since this check is initiating communication with the frontboard
service which will be blocked, this check should be moved to the UI process. In the UI process, this is checked when
starting a new WebContent process, and sent to the WebContent process as part of the process creation parameters. The
WebContent is storing the received value.

API test: WebKit.PictureInPictureSupport

* platform/PictureInPictureSupport.h:
* platform/ios/VideoFullscreenInterfaceAVKit.mm:
(WebCore::setSupportsPictureInPicture):
(WebCore::supportsPictureInPicture):
* testing/Internals.cpp:
(WebCore::Internals::supportsPictureInPicture):
* testing/Internals.h:
* testing/Internals.idl:

Source/WebKit:

When starting a new WebContent process, check if PIP is supported, and pass the results to the newly created
WebContent process.

* Shared/WebProcessCreationParameters.cpp:
(WebKit::WebProcessCreationParameters::encode const):
(WebKit::WebProcessCreationParameters::decode):
* Shared/WebProcessCreationParameters.h:
* UIProcess/Cocoa/WebProcessPoolCocoa.mm:
(WebKit::WebProcessPool::platformInitializeWebProcess):
* WebProcess/cocoa/WebProcessCocoa.mm:
(WebKit::WebProcess::platformInitializeWebProcess):

Tools:

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/Tests/WebKit/PictureInPictureSupport.mm: Added.
(TEST):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@256531 268f45cc-cd09-0410-ab3c-d52691b4dbfc