[First paint] Let's check visually-non-empty state in FrameView::styleAndRenderTreeDidChange https://bugs.webkit.org/show_bug.cgi?id=208013 <rdar://problem/59633427>
Reviewed by Antti Koivisto.

This patch completely decouples m_contentQualifiesAsVisuallyNonEmpty and m_firstVisuallyNonEmptyLayoutMilestoneIsPending.
Now FrameView can be in the state where m_contentQualifiesAsVisuallyNonEmpty is true while
m_firstVisuallyNonEmptyLayoutMilestoneIsPending is false (which is a perfectly valid state to be in).
This will enable us to delay optional layouts after style recalcs when the content is still visually empty.
(If a FrameView::isVisuallyNonEmpty() caller regresses after this change, it means that the caller is interested in
whether the non-empty milestone has been issued or not and it needs to switch over to a different API.)

* page/FrameView.cpp:
(WebCore::FrameView::styleAndRenderTreeDidChange):
(WebCore::FrameView::loadProgressingStatusChanged):
(WebCore::FrameView::fireLayoutRelatedMilestonesIfNeeded):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@257081 268f45cc-cd09-0410-ab3c-d52691b4dbfc