fast/forms/ios/click-should-not-suppress-misspelling.html fails on iOS 13.4 beta https://bugs.webkit.org/show_bug.cgi?id=208086 <rdar://problem/59631501>
Reviewed by Tim Horton.

In the iOS 13.4 beta, tapping near the caret rect in an editable text field causes the callout bar to toggle
visibility instead of triggering word-granularity spellchecking and changing the selection. This test in
particular dispatches two taps: the first at (100, 100), and the second at (300, 100) in content view
coordinates. However, since the page does not have a viewport, the entire page is scaled down on iPhone, such
that the second tap at (300, 100) ends up being very close to the caret rect.

To fix this, we simply make the text field take up 100% of the viewport width, and make the viewport use
device-width with an initial scale of 1. This ensures that the second tap will be somewhere near the end of the
misspelled word, which triggers spellchecking as intended by the test instead of just showing the callout bar.

* fast/forms/ios/click-should-not-suppress-misspelling-expected.txt:
* fast/forms/ios/click-should-not-suppress-misspelling.html:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@257223 268f45cc-cd09-0410-ab3c-d52691b4dbfc