Web Inspector: set both the page's and the window's appearance when modifying the frontend forced appearance in the case that the frontend is docked https://bugs.webkit.org/show_bug.cgi?id=208557
Reviewed by Timothy Hatcher.

* inspector/InspectorFrontendHost.cpp:
(WebCore::InspectorFrontendHost::setForcedAppearance):
`InspectorFrontendClient::setForcedAppearance` only has an effect when Web Inspector is
detached/undocked, meaning that changing the Appearance in the Settings Tab when docked has
no effect. In addition to notifying the `InspectorFrontendClient`, we should also call
`Page::setUseDarkAppearanceOverride` so that if Web Inspector is docked, the apperance
changes. The Web Inspector frontend page's apperance would normally change to match the
appearance of the window, but when Web Inspector is docked there is no window.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@257833 268f45cc-cd09-0410-ab3c-d52691b4dbfc