Call globalObjectIsAvailableForFrame before evaluating JavaScript in newly created worlds https://bugs.webkit.org/show_bug.cgi?id=208615 <rdar://problem/59406743>
Reviewed by Alex Christensen.

Source/WebKit:

globalObjectIsAvailableForFrame is often used for initialization code for a content world.
If we evaluate JavaScript in a content world as the first thing we do in that world, 
globalObjectIsAvailableForFrame hadn't been called, so things were not initialized.
So call it when evaluating JavaScript if we create the world to evaluate in.

Covered by new API tests.

* WebProcess/UserContent/WebUserContentController.cpp:
(WebKit::WebUserContentController::addContentWorld):
* WebProcess/UserContent/WebUserContentController.h:
* WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::runJavaScriptInFrameInScriptWorld):

Tools:

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/Tests/WebKitCocoa/ContentWorldPlugIn.mm: Added.
(-[ContentWorldPlugIn webProcessPlugIn:didCreateBrowserContextController:]):
(-[ContentWorldPlugIn webProcessPlugInBrowserContextController:globalObjectIsAvailableForFrame:inScriptWorld:]):
* TestWebKitAPI/Tests/WebKitCocoa/WKWebViewEvaluateJavaScript.mm:
(TEST):
Also make a test written in bug 206310 actually run code.



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@257892 268f45cc-cd09-0410-ab3c-d52691b4dbfc