Crash accessing AXIsolatedObject::m_childrenIDS from removeSubtree on the main thread. https://bugs.webkit.org/show_bug.cgi?id=208728
Reviewed by Chris Fleizach.

AXIsolatedTree::removeSubtree was accessing AXIsolatedObject::m_childrenIDs
on the main thread to remove all descendants recursively. But the lock
had to be unlocked and locked again on each iteration, creating problems
if the secondary thread modifies the children in between iterations.
The solution in this patch is to eliminate removeSubtree, and make
removeNode and applyPendingChanges to remove all descendants.

* accessibility/AXObjectCache.cpp:
(WebCore::AXObjectCache::updateIsolatedTree):
* accessibility/isolatedtree/AXIsolatedTree.cpp:
(WebCore::AXIsolatedTree::appendNodeChanges): Renamed local vars to make it clearer.
(WebCore::AXIsolatedTree::applyPendingChanges):
(WebCore::AXIsolatedTree::removeSubtree): Deleted.
* accessibility/isolatedtree/AXIsolatedTree.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@258034 268f45cc-cd09-0410-ab3c-d52691b4dbfc